Remove deprecated option
